transfer ascii10 to new main source file
